refactor(settings): isolate settings-related organisms to features/settings (#1455)
