+++ eslint.config.mjs
@@ -51,6 +51,7 @@ export default ts.config(
           mode: "file",
           pattern: [
             "src/components/**/*",
+            "src/features/**/*",
             "src/pages/**/*",
             "src/tabs/**/*",
             "src/options.tsx",
@@ -197,13 +198,13 @@ export default ts.config(
       "src/components/organisms/CardDetailView.tsx",
       "src/components/organisms/Workbench.tsx",
       "src/components/organisms/InteractiveTutorial.tsx",
-      "src/components/organisms/WebLlmSettingsSection.tsx",
+      "src/features/settings/components/WebLlmSettingsSection.tsx",
       "src/components/organisms/SimpleWorkbenchModal.tsx",
       "src/components/organisms/SlotVariablesSection.tsx",
-      "src/components/organisms/StorageManagerSection.tsx",
+      "src/features/settings/components/StorageManagerSection.tsx",
       "src/components/organisms/CategoryManagerModal.tsx",
-      "src/components/organisms/CloudSyncSection.tsx",
-      "src/components/organisms/LocalBackupSection.tsx",
+      "src/features/settings/components/CloudSyncSection.tsx",
+      "src/features/settings/components/LocalBackupSection.tsx",
       "src/components/organisms/Cauldron.tsx",
       "src/components/molecules/AssociatedImageGallery.tsx",
       "src/components/molecules/TagEditor.tsx",
@@ -307,7 +308,7 @@ export default ts.config(
       "src/components/molecules/WebGpuWarning.tsx",
       "src/components/organisms/AiRecipeAdviceSection.tsx",
       "src/components/organisms/AiStyleAnalysisSection.tsx",
-      "src/components/organisms/BrandLogoSettingsSection.tsx",
+      "src/features/settings/components/BrandLogoSettingsSection.tsx",
       "src/components/organisms/CardDetailView/SubSections.tsx",
       "src/components/organisms/CardsGrid.tsx",
       "src/components/organisms/CategoryFilters.tsx",
@@ -316,8 +317,8 @@ export default ts.config(
       "src/components/organisms/CategoryManagerModal/CategoryList.tsx",
       "src/components/organisms/CategoryManagerModal/CategoryModalHeader.tsx",
       "src/components/organisms/CategoryManagerModal/CoverSettingsFields.tsx",
-      "src/components/organisms/CloudSyncSection.tsx",
-      "src/components/organisms/DangerZoneSection.tsx",
+      "src/features/settings/components/CloudSyncSection.tsx",
+      "src/features/settings/components/DangerZoneSection.tsx",
       "src/components/organisms/DbErrorOverlay.tsx",
       "src/components/organisms/EmptyState.tsx",
       "src/components/organisms/EvolutionSuccessModal.tsx",
@@ -327,7 +328,7 @@ export default ts.config(
       "src/components/organisms/HandBarScrollButton.tsx",
       "src/components/organisms/LibraryFilterAccordion.tsx",
       "src/components/organisms/LibrarySearchBar.tsx",
-      "src/components/organisms/LocalBackupSection.tsx",
+      "src/features/settings/components/LocalBackupSection.tsx",
       "src/components/organisms/MergeStackMaterialList.tsx",
       "src/components/organisms/MintingView.tsx",
       "src/components/organisms/MintingView/CardIdentitySubSections.tsx",
@@ -344,11 +345,11 @@ export default ts.config(
       "src/components/molecules/simple-minting/KeywordSuggestions.tsx",
       "src/components/molecules/simple-minting/ThumbnailPreview.tsx",
       "src/components/organisms/SimpleWorkbenchModal.tsx",
-      "src/components/organisms/StorageManagerSection.tsx",
+      "src/features/settings/components/StorageManagerSection.tsx",
       "src/components/organisms/SyncStatusMessage.tsx",
       "src/components/organisms/TipsBar.tsx",
       "src/components/organisms/TutorialTooltip.tsx",
-      "src/components/organisms/UiPreferencesSection.tsx",
+      "src/features/settings/components/UiPreferencesSection.tsx",
       "src/components/organisms/WorkbenchCard.tsx",
       "src/components/organisms/WorkbenchHeader.tsx",
       "src/components/templates/SidePanelDebugLogs.tsx",
