+++ src/features/settings/components/LocalBackupSection.tsx
@@ -8,7 +8,7 @@ import {
 } from "lucide-react"
 import React from "react"
 
-import { HelpTooltip } from "../atoms/HelpTooltip"
+import { HelpTooltip } from "../../../components/atoms/HelpTooltip"
 
 interface LocalBackupSectionProps {
   fileInputRef: React.RefObject<HTMLInputElement | null>
