+++ src/features/settings/components/StorageManagerSection.tsx
@@ -1,7 +1,7 @@
 import { AlertTriangle, Database, Trash2 } from "lucide-react"
 import React from "react"
 
-import { HelpTooltip } from "../atoms/HelpTooltip"
+import { HelpTooltip } from "../../../components/atoms/HelpTooltip"
 
 interface StorageManagerSectionProps {
   estimate: any
